+++ tools/model_converters/llava-delta2mmpre.py
@@ -9,23 +9,21 @@ from huggingface_hub import snapshot_download
 from transformers.modeling_utils import load_state_dict
 
 prog_description = """\
-Merge Llava delta weights and original weights,
-and save as MMPreTrain checkpoint.
+Convert Llava weights and original weights.
 """
 
 
 def parse_args():
     parser = argparse.ArgumentParser(description=prog_description)
-    parser.add_argument(
-        'src_path', type=str, help='The original checkpoint dir')
-    parser.add_argument(
-        'delta_path', type=str, help='The delta checkpoint dir')
-    parser.add_argument('dst_path', type=str, help='The saved checkpoint path')
+    parser.add_argument('src', type=str, help='The original checkpoint dir')
+    parser.add_argument('dst', type=str, help='The saved checkpoint path')
+    parser.add_argument('--delta', type=str, help='The delta checkpoint dir')
     args = parser.parse_args()
     return args
 
 
 def load_checkpoint(path: Path):
+    path = Path(path)
     if path.is_file():
         return torch.load(path)
 
@@ -41,19 +39,23 @@ def load_checkpoint(path: Path):
 def main():
     args = parse_args()
 
-    if Path(args.src_path).exists():
-        src_path = Path(args.src_path)
+    if Path(args.src).exists():
+        src_path = args.src
     else:
-        src_path = Path(snapshot_download(args.src_path))
+        src_path = snapshot_download(
+            args.src, allow_patterns='pytorch_model*.bin')
     src_state_dict = load_checkpoint(src_path)
 
-    if Path(args.delta_path).exists():
-        delta_path = Path(args.delta_path)
+    if args.delta is None:
+        delta_state_dict = {}
+    elif Path(args.delta).exists():
+        delta_state_dict = load_checkpoint(args.delta)
     else:
-        delta_path = Path(snapshot_download(args.delta_path))
-    delta_state_dict = load_checkpoint(delta_path)
+        delta_path = snapshot_download(
+            args.delta, allow_patterns='pytorch_model*.bin')
+        delta_state_dict = load_checkpoint(delta_path)
 
-    merged_state_dict = OrderedDict()
+    new_state_dict = OrderedDict()
     for k, v in src_state_dict.items():
         if k in delta_state_dict:
             delta_v = delta_state_dict.pop(k)
@@ -63,12 +65,13 @@ def main():
                 v = delta_v
             else:
                 v += delta_v
-        merged_state_dict['model.lang_encoder.' + k] = v
+        if 'rotary_emb.inv_freq' not in k:
+            new_state_dict['model.lang_encoder.' + k] = v
 
     for k, v in delta_state_dict.items():
-        merged_state_dict['model.lang_encoder.' + k] = v
+        new_state_dict['model.lang_encoder.' + k] = v
 
-    torch.save(merged_state_dict, args.dst_path)
+    torch.save(new_state_dict, args.dst)
     print('Done!!')